move Getter/Setter for scheduleStopper to Entity.java
